Extracted gameobjects from actor base, corrected some messages and removed duplicated code
